activated -Wno-double-promotion

+++ src/lib/utils/sqlite_wrapper.cpp
@@ -326,7 +326,7 @@ void SQLiteWrapper::create_sqlite_table(const Table& table, const std::string& t
               break;
             case DataType::Float:
               sqlite3_bind_return_code =
-                  sqlite3_bind_double(insert_into_statement, sqlite_column_id, boost::get<float>(value));
+                  sqlite3_bind_double(insert_into_statement, sqlite_column_id, static_cast<double>(boost::get<float>(value)));
               break;
             case DataType::Double:
               sqlite3_bind_return_code =
